requirements.txt
langgraph>=0.2
langchain-anthropic>=0.1
anthropic>=0.30
fastapi
fastapi-cache2
uvicorn
httpx
supabase>=2.4
twilio
deepgram-sdk>=3
clerk-backend-api
python-jose[cryptography]
python-dotenv
pydantic>=2
pydantic-settings
rich
apscheduler
langgraph-checkpoint-postgres
psycopg[binary]
